requirements.txt
selenium
webdriver-manager
selenium_stealth
bs4
pandas
python-docx
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=0.24.0
transformers>=4.12.0
sentence-transformers>=2.2.0
torch>=1.10.0
google-generativeai 
ragas
langchain
langchain-google-genai 
langchain-community
ollama